Reorganize sensors code for ESP32 Azure IoT Kit sample
