--- demos/projects/ESPRESSIF/aziotkit/components/azure-iot-kit-sensors/src/sensors/motor_l298n.h
@@ -1,55 +1,0 @@
-/* brushed dc motor control example
-   This example code is in the Public Domain (or CC0 licensed, at your option.)
-   Unless required by applicable law or agreed to in writing, this
-   software is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied.
-*/
-
-/*
- * This example will show you how to use MCPWM module to control brushed dc motor.
- * This code is tested with L298 motor driver.
- * User may need to make changes according to the motor driver they use.
-*/
-#ifndef _MOTOR_L298N_H_
-#define _MOTOR_L298N_H_
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-#include <stdio.h>
-
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include "esp_attr.h"
-
-#include "driver/mcpwm.h"
-#include "soc/mcpwm_reg.h"
-#include "soc/mcpwm_struct.h"
-
-#define GPIO_PWM0A_OUT 12   //Set GPIO 15 as PWM0A
-#define GPIO_PWM0B_OUT 13   //Set GPIO 16 as PWM0B
-
-void mcpwm_example_gpio_initialize();
-
-/**
- * @brief motor moves in forward direction, with duty cycle = duty %
- */
-void brushed_motor_forward(mcpwm_unit_t mcpwm_num, mcpwm_timer_t timer_num , float duty_cycle);
-
-/**
- * @brief motor moves in backward direction, with duty cycle = duty %
- */
-void brushed_motor_backward(mcpwm_unit_t mcpwm_num, mcpwm_timer_t timer_num , float duty_cycle);
-
-/**
- * @brief motor stop
- */
-void brushed_motor_stop(mcpwm_unit_t mcpwm_num, mcpwm_timer_t timer_num);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
